[Round2][Django Model] 4-19. 접근 제어 믹스인 분리
